+++ src/components/chromeExtensions/MainPopup/styles.js
