+++ src/components/chromeExtensions/PopupError/styles.js
