+++ src/components/chromeExtensions/WelcomePopup/index.js
