+++ src/components/chromeExtensions/WelcomePopup/styles.js
