+++ src/components/shared/PasswordInfoModal/styles.js
