+++ src/components/shared/ShowModal/styles.js
